I did some changes in MissingNumberInArray question

+++ MissingNumberInArray.java
@@ -1,12 +1,12 @@
 public class MissingNumberInArray {
     public static int missing(int arr[]) {
-        int n = arr.length + 1;
-        int sum = n*(n+1)/2;
-        int arraySum = 0;
+        long n = arr.length + 1;
+        long sum = n*(n+1)/2;
+        long arraySum = 0;
         for (int i = 0; i < arr.length; i++) {
             arraySum += arr[i];
         }
-        return sum - arraySum;
+        return (int)(sum - arraySum);
     }
     public static void main(String[] args) {
         int arr[] = {1, 2, 3, 4, 6, 7};
